Predicates

Apply a predicate to check if all / any elements in a collection return true. Could be implemented as a reduction, but is better optimised with stopping the search once a false / true is found.

Other names: not often implemented standalone on GPUs, typically included as part of a reduction.

AcceleratedKernels.any — Function

any(
    pred,
    v::AbstractArray;

    # CPU settings
    max_tasks=Threads.nthreads(),
    min_elems=1,

    # GPU settings
    block_size::Int=256,
    cooperative::Bool=true,

    # GPU settings passed to mapreduce, only used if cooperative=false
    temp::Union{Nothing, AbstractArray}=nothing,
    switch_below::Int=0,
)

Check if any element of v satisfies the predicate pred (i.e. some pred(v[i]) == true). Optimised differently to mapreduce due to shortcircuiting behaviour of booleans.

Other names: not often implemented standalone on GPUs, typically included as part of a reduction.

On the CPU, parallelisation is only worth it for large arrays, relatively expensive predicates, and/or rare occurrence of true; use max_tasks and min_elems to only use parallelism when worth it in your application. When only one thread is needed, there is no overhead.

On the GPU, the cooperative flag controls whether to use an optimised implementation which works by concurrent writing to a global flag; there is only one platform we are aware of (old Intel UHD 620 integrated graphics cards) where such writes hang. In such cases, set cooperative=false to use a mapreduce implementation, for which you can also use temp and switch_below.

Examples

import AcceleratedKernels as AK
using CUDA

v = CuArray(rand(Float32, 100_000))
AK.any(x -> x < 1, v)

AcceleratedKernels.all — Function

all(
    pred,
    v::AbstractGPUArray;

    # CPU settings
    max_tasks=Threads.nthreads(),
    min_elems=1,

    # GPU settings
    block_size::Int=256,
    cooperative::Bool=true,

    # GPU settings passed to mapreduce, only used if cooperative=false
    temp::Union{Nothing, AbstractArray}=nothing,
    switch_below::Int=0,
)

Check if all elements of v satisfy the predicate pred (i.e. all pred(v[i]) == true). Optimised differently to mapreduce due to shortcircuiting behaviour of booleans.

Other names: not often implemented standalone on GPUs, typically included as part of a reduction.

On the CPU, parallelisation is only worth it for large arrays, relatively expensive predicates, and/or rare occurrence of false; use max_tasks and min_elems to only use parallelism when worth it in your application. When only one thread is needed, there is no overhead.

On the GPU, the cooperative flag controls whether to use an optimised implementation which works by concurrent writing to a global flag; there is only one platform we are aware of (old Intel UHD 620 integrated graphics cards) where such writes hang. In such cases, set cooperative=false to use a mapreduce implementation, for which you can also use temp and switch_below.

Examples

```julia import AcceleratedKernels as AK using Metal

v = MtlArray(rand(Float32, 100_000)) AK.all(x -> x > 0, v) ````

Note on the cooperative keyword: some older platforms crash when multiple threads write to the same memory location in a global array (e.g. old Intel Graphics); if all threads were to write the same value, it is well-defined on others (e.g. CUDA F4.2 says "If a non-atomic instruction executed by a warp writes to the same location in global memory for more than one of the threads of the warp, only one thread performs a write and which thread does it is undefined."). This "cooperative" thread behaviour allows for a faster implementation; if you have a platform - the only one I know is Intel UHD Graphics - that crashes, set cooperative=false to use a safer mapreduce-based implementation.